refactor i18n structure
