+++ src/boot/commonSetup.ts
@@ -1,6 +1,6 @@
 import { App } from "vue";
 import { createPinia } from "pinia";
-import { setupI18n } from "./i18n";
+import { setupI18n } from "@/i18n";
 
 import "@/assets/styles/index.css";
