--- src/locales/pt-BR.json
@@ -1,7 +1,0 @@
-{
-  "mainMenu": {
-    "newWallet": "Nova carteira",
-    "settings": "Configurações",
-    "about": "Sobre"
-  }
-}
