+++ src/stores/appStore.ts
@@ -6,7 +6,6 @@ import { useColorMode } from "@vueuse/core";
 import AES from "crypto-js/aes";
 import { Locale } from "vue-i18n";
 import { useRouter } from "vue-router";
-import { setLocale } from "@/boot/i18n";
 import HdKey from "@/chains/ergo/hdKey";
 import { graphQLService } from "@/chains/ergo/services/graphQlService";
 import { hdKeyPool } from "@/common/objectPool";
@@ -16,6 +15,7 @@ import { DEFAULT_SETTINGS } from "@/constants/settings";
 import { utxosDbService } from "@/database/utxosDbService";
 import { WalletPatch, walletsDbService } from "@/database/walletsDbService";
 import { sendBackendServerUrl } from "@/extension/connector/rpc/uiRpcHandlers";
+import { setLocale } from "@/i18n";
 import { IDbWallet, NotNullId } from "@/types/database";
 import { Network, WalletType } from "@/types/internal";
 import { useChainStore } from "./chainStore";
