+++ src/types/d.ts/i18n.d.ts
@@ -1,7 +1,7 @@
 // Interfaces need to imported first to be extended
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 import type { DefineDateTimeFormat, DefineLocaleMessage, DefineNumberFormat } from "vue-i18n";
-import type { MessageSchema, SUPPORTED_LOCALES } from "@/boot/i18n";
+import type { MessageSchema, SUPPORTED_LOCALES } from "@/i18n/i18n";
 
 declare module "vue-i18n" {
   // define the locale messages schema
